+++ src/models/graphdata.h
@@ -39,8 +39,8 @@ public:
     quint16 registerAddress() const;
     void setRegisterAddress(const quint16 &registerAddress);
 
-    quint16 bitmask() const;
-    void setBitmask(const quint16 &bitmask);
+    quint32 bitmask() const;
+    void setBitmask(const quint32 &bitmask);
 
     qint32 shift() const;
     void setShift(const qint32 &shift);
@@ -62,7 +62,7 @@ private:
     double _multiplyFactor;
     double _divideFactor;
     quint16 _registerAddress;
-    quint16 _bitmask;
+    quint32 _bitmask;
     qint32 _shift;
     quint8 _connectionId;
